allow TE mint at position ops
