+++ programs/whirlpool/src/instructions/collect_fees.rs
@@ -1,9 +1,10 @@
 use anchor_lang::prelude::*;
 use anchor_spl::token::{self, Token, TokenAccount};
+use anchor_spl::token_interface::TokenAccount as TokenAccountInterface;
 
 use crate::{
     state::*,
-    util::{transfer_from_vault_to_owner, verify_position_authority},
+    util::{transfer_from_vault_to_owner, verify_position_authority_2022},
 };
 
 #[derive(Accounts)]
@@ -18,7 +19,7 @@ pub struct CollectFees<'info> {
         constraint = position_token_account.mint == position.position_mint,
         constraint = position_token_account.amount == 1
     )]
-    pub position_token_account: Box<Account<'info, TokenAccount>>,
+    pub position_token_account: Box<InterfaceAccount<'info, TokenAccountInterface>>,
 
     #[account(mut, constraint = token_owner_account_a.mint == whirlpool.token_mint_a)]
     pub token_owner_account_a: Box<Account<'info, TokenAccount>>,
@@ -35,7 +36,7 @@ pub struct CollectFees<'info> {
 }
 
 pub fn handler(ctx: Context<CollectFees>) -> Result<()> {
-    verify_position_authority(
+    verify_position_authority_2022(
         &ctx.accounts.position_token_account,
         &ctx.accounts.position_authority,
     )?;
